feat: update README and execute install only when Gemfile exists

+++ docker-entrypoint.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 set -e
 
-bundle install --retry 5 --jobs 20
+if [ -f Gemfile ]; then
+  bundle install --retry 5 --jobs 20
+fi
 
 exec "$@"
